refactor: rewrite CopyProperties helper
